+++ ydb/core/http_proxy/auth_factory.cpp
@@ -1,5 +1,6 @@
 #include "auth_factory.h"
 #include "http_req.h"
+#include "auth_actors.h"
 #include <ydb/core/base/feature_flags.h>
 #include <ydb/core/http_proxy/http_service.h>
 #include <ydb/core/http_proxy/http_req.h>
@@ -98,4 +99,3 @@ NActors::IActor* TIamAuthFactory::CreateAuthActor(const TActorId sender, THttpRe
 
 
 }
-
