+++ ydb/core/http_proxy/ut/datastreams_fixture/datastreams_fixture.cpp
@@ -1,5 +1,7 @@
 #include "datastreams_fixture.h"
 
+#include <ydb/core/http_proxy/auth_actors.h>
+
 #include <util/system/env.h>
 
 #include <thread>
